requirements.txt
flask
PyMySQL
thread6
mysql-connector-python
smbus
paho-mqtt